app/requirements.txt
luma.oled==3.*
Pillow==10.*
